backend/business-service-ms/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
python-dotenv>=1.0.0
django-cors-headers>=4.2.0
drf-yasg>=1.21.7
psycopg2>=2.9.9
PyJWT>=2.8.0
requests>=2.31.0
django-filter>=23.3